backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0

# Ecosistema LangChain 0.3
langchain==0.3.27
langchain-community==0.3.27
langgraph==0.2.27
langchain-google-genai==2.1.9

# Vector store + utilidades
faiss-cpu==1.8.0.post1
beautifulsoup4==4.12.3
httpx==0.27.0
python-dotenv==1.0.1

pinecone==5.4.2
langchain-pinecone==0.2.2
aiohttp>=3.10,<3.11


